+++ integration/src/test/scala/co/topl/bridge/BridgeIntegrationSpec.scala
@@ -2,8 +2,6 @@ package co.topl.bridge
 
 import cats.effect.IO
 import co.topl.shared.BridgeContants
-import co.topl.shared.ConfirmDepositRequest
-import co.topl.shared.ConfirmDepositResponse
 import co.topl.shared.ConfirmRedemptionRequest
 import co.topl.shared.ConfirmRedemptionResponse
 import co.topl.shared.StartPeginSessionRequest
@@ -20,6 +18,9 @@ import org.http4s.ember.client._
 import org.http4s.headers.`Content-Type`
 
 import scala.concurrent.duration._
+import org.checkerframework.checker.units.qual.s
+import co.topl.shared.MintingStatusRequest
+import co.topl.shared.MintingStatusResponse
 
 class BridgeIntegrationSpec extends CatsEffectSuite {
 
@@ -33,6 +34,8 @@ class BridgeIntegrationSpec extends CatsEffectSuite {
     "bitcoin-cli",
     "-regtest",
     "-named",
+    "-rpcuser=bitcoin",
+    "-rpcpassword=password",
     "createwallet",
     "wallet_name=testwallet"
   )
@@ -40,6 +43,8 @@ class BridgeIntegrationSpec extends CatsEffectSuite {
     "exec",
     "bitcoin",
     "bitcoin-cli",
+    "-rpcuser=bitcoin",
+    "-rpcpassword=password",
     "-regtest",
     "-rpcwallet=testwallet",
     "getnewaddress"
@@ -49,6 +54,8 @@ class BridgeIntegrationSpec extends CatsEffectSuite {
     "bitcoin",
     "bitcoin-cli",
     "-regtest",
+    "-rpcuser=bitcoin",
+    "-rpcpassword=password",
     "generatetoaddress",
     blocks.toString,
     address
@@ -58,6 +65,8 @@ class BridgeIntegrationSpec extends CatsEffectSuite {
     "bitcoin",
     "bitcoin-cli",
     "-regtest",
+    "-rpcuser=bitcoin",
+    "-rpcpassword=password",
     "generatetoaddress",
     "101",
     address
@@ -67,6 +76,8 @@ class BridgeIntegrationSpec extends CatsEffectSuite {
     "bitcoin",
     "bitcoin-cli",
     "-regtest",
+    "-rpcuser=bitcoin",
+    "-rpcpassword=password",
     "-rpcwallet=testwallet",
     "signrawtransactionwithwallet",
     tx
@@ -76,6 +87,8 @@ class BridgeIntegrationSpec extends CatsEffectSuite {
     "bitcoin",
     "bitcoin-cli",
     "-regtest",
+    "-rpcuser=bitcoin",
+    "-rpcpassword=password",
     "sendrawtransaction",
     signedTx
   )
@@ -84,6 +97,8 @@ class BridgeIntegrationSpec extends CatsEffectSuite {
     "exec",
     "bitcoin",
     "bitcoin-cli",
+    "-rpcuser=bitcoin",
+    "-rpcpassword=password",
     "-regtest",
     "-rpcwallet=testwallet",
     "listunspent"
@@ -173,21 +188,21 @@ class BridgeIntegrationSpec extends CatsEffectSuite {
     implicit val syncWalletRequestDecoder
         : EntityEncoder[IO, SyncWalletRequest] =
       jsonEncoderOf[IO, SyncWalletRequest]
+    implicit val mintingStatusRequesEncoder
+        : EntityEncoder[IO, MintingStatusRequest] =
+      jsonEncoderOf[IO, MintingStatusRequest]
     implicit val startSessionResponse
         : EntityDecoder[IO, StartPeginSessionResponse] =
       jsonOf[IO, StartPeginSessionResponse]
+    implicit val MintingStatusResponseDecoder
+        : EntityDecoder[IO, MintingStatusResponse] =
+      jsonOf[IO, MintingStatusResponse]
     implicit val confirmRedemptionRequestDecoder
         : EntityEncoder[IO, ConfirmRedemptionRequest] =
       jsonEncoderOf[IO, ConfirmRedemptionRequest]
     implicit val confirmRedemptionResponse
         : EntityDecoder[IO, ConfirmRedemptionResponse] =
       jsonOf[IO, ConfirmRedemptionResponse]
-    implicit val confirmDepositRequestEncoder
-        : EntityEncoder[IO, ConfirmDepositRequest] =
-      jsonEncoderOf[IO, ConfirmDepositRequest]
-    implicit val confirmDepositRequestDecoder
-        : EntityDecoder[IO, ConfirmDepositResponse] =
-      jsonOf[IO, ConfirmDepositResponse]
     assertIO(
       for {
         createWalletOut <- process
@@ -238,27 +253,7 @@ class BridgeIntegrationSpec extends CatsEffectSuite {
               )
             )
           })
-        syncWalletResponse <- EmberClientBuilder
-          .default[IO]
-          .build
-          .use({ client =>
-            client.expect[String](
-              Request[IO](
-                method = Method.POST,
-                Uri
-                  .fromString("http://127.0.0.1:4000/api/sync-wallet")
-                  .toOption
-                  .get
-              ).withContentType(
-                `Content-Type`.apply(MediaType.application.json)
-              ).withEntity(
-                SyncWalletRequest(
-                  "secret"
-                )
-              )
-            )
-          })
-        _ <- IO.println("syncWalletResponse: " + syncWalletResponse)
+        _ <- IO.println("Escrow address: " + startSessionResponse.escrowAddress)
         bitcoinTx <- process
           .ProcessBuilder(
             DOCKER_CMD,
@@ -278,79 +273,36 @@ class BridgeIntegrationSpec extends CatsEffectSuite {
           .ProcessBuilder(DOCKER_CMD, sendTransaction(signedTxHex): _*)
           .spawn[IO]
           .use(getText)
-        genusQueryresult <- getCurrentUtxos.use(getText)
-        txId = genusQueryresult
-          .split("\n")
-          .filter(_.endsWith("#1"))
-          .head
-          .split(":")
-          .map(_.trim)
-          .tail
-          .head
-          .split("#")
-          .head
-        confirmDepositResponse <- EmberClientBuilder
-          .default[IO]
-          .build
-          .use({ client =>
-            client.expect[ConfirmDepositResponse](
-              Request[IO](
-                method = Method.POST,
-                Uri
-                  .fromString("http://127.0.0.1:4000/api/confirm-deposit-btc")
-                  .toOption
-                  .get
-              ).withContentType(
-                `Content-Type`.apply(MediaType.application.json)
-              ).withEntity(
-                ConfirmDepositRequest(
-                  startSessionResponse.sessionID,
-                  4999000000L
-                )
-              )
-            )
-          })
-        _ <- IO.println("confirmDepositResponse: " + confirmDepositResponse)
-        _ <- IO.sleep(10.seconds)
-        _ <- getCurrentUtxosFromAddress(confirmDepositResponse.redeemAddress)
-          .use(_.exitValue)
-          .iterateUntil(_ == 0)
-        _ <- assertIOBoolean(
-          getCurrentUtxosFromAddress(confirmDepositResponse.redeemAddress)
-            .use(getText)
-            .map { x =>
-              x.contains("Asset")
-            }
-        )
         _ <- process
-          .ProcessBuilder(DOCKER_CMD, generateToAddress(10, newAddress): _*)
+          .ProcessBuilder(DOCKER_CMD, generateToAddress(6, newAddress): _*)
           .spawn[IO]
           .use(_.exitValue)
-        confirmRedemptionResponse <- EmberClientBuilder
+        _ <- EmberClientBuilder
           .default[IO]
           .build
           .use({ client =>
-            client.expect[ConfirmRedemptionResponse](
-              Request[IO](
-                method = Method.POST,
-                Uri
-                  .fromString("http://127.0.0.1:4000/api/confirm-redemption")
-                  .toOption
-                  .get
-              ).withContentType(
-                `Content-Type`.apply(MediaType.application.json)
-              ).withEntity(
-                ConfirmRedemptionRequest(
-                  startSessionResponse.sessionID,
-                  sentTxId,
-                  0,
-                  2,
-                  4999000000L,
-                  "secret"
+            (IO.println("Requesting..") >> client
+              .expect[MintingStatusResponse](
+                Request[IO](
+                  method = Method.POST,
+                  Uri
+                    .fromString(
+                      "http://127.0.0.1:4000/api/" + BridgeContants.TOPL_MINTING_STATUS
+                    )
+                    .toOption
+                    .get
+                ).withContentType(
+                  `Content-Type`.apply(MediaType.application.json)
+                ).withEntity(
+                  MintingStatusRequest(startSessionResponse.sessionID)
                 )
               )
-            )
+              .flatMap(x => IO.println(x.mintingStatus) >> IO.sleep(5.second) >> IO.pure(x)))
+              .iterateUntil(
+                _.mintingStatus == "PeginSessionWaitingForRedemption"
+              )
           })
+
       } yield (),
       ()
     )