+++ topl-btc-bridge/src/main/scala/co/topl/bridge/BridgeParamsDescriptor.scala
@@ -70,7 +70,31 @@ trait BridgeParamsDescriptor {
         .action((x, c) => c.copy(toplHost = x))
         .text("The host of the Topl node. (mandatory)")
         .validate(x =>
-          if (x.trim().isEmpty) failure("Host may not be empty") else success
+          if (x.trim().isEmpty) failure("Topl node host may not be empty")
+          else success
+        ),
+      opt[String]("btc-url")
+        .action((x, c) => c.copy(btcUrl = x))
+        .text("The url of the Bitcoin node. (mandatory)")
+        .validate(x =>
+          if (x.trim().isEmpty) failure("Bitcoin node url may not be empty")
+          else success
+        ),
+      opt[String]("btc-user")
+        .action((x, c) => c.copy(btcUser = x))
+        .text("The username for the Bitcoin node. (mandatory)")
+        .validate(x =>
+          if (x.trim().isEmpty)
+            failure("Bitcoin node username may not be empty")
+          else success
+        ),
+      opt[String]("btc-password")
+        .action((x, c) => c.copy(btcPassword = x))
+        .text("The password for the Bitcoin node. (mandatory)")
+        .validate(x =>
+          if (x.trim().isEmpty)
+            failure("Bitcoin node password may not be empty")
+          else success
         ),
       opt[Int]("topl-port")
         .action((x, c) => c.copy(toplPort = x))
@@ -78,6 +102,20 @@ trait BridgeParamsDescriptor {
         .validate(x =>
           if (x >= 0 && x <= 65536) success
           else failure("Port must be between 0 and 65536")
+        ),
+      opt[Int]("zmq-port")
+        .action((x, c) => c.copy(zmqPort = x))
+        .text("Port for ZMQ. (optional)")
+        .validate(x =>
+          if (x >= 0 && x <= 65536) success
+          else failure("Port must be between 0 and 65536")
+        ),
+      opt[String]("zmq-host")
+        .action((x, c) => c.copy(zmqHost = x))
+        .text("Host for ZMQ. (optional)")
+        .validate(x =>
+          if (x.trim().isEmpty) failure("ZMQ host may not be empty")
+          else success
         ),
       opt[Boolean]("topl-secure")
         .action((x, c) => c.copy(toplSecureConnection = x))