+++ topl-btc-bridge/src/main/scala/co/topl/bridge/Main.scala
@@ -3,15 +3,20 @@ package co.topl.bridge
 import cats.effect.ExitCode
 import cats.effect.IO
 import cats.effect.IOApp
+import cats.effect.kernel.Ref
+import co.topl.brambl.monitoring.BitcoinMonitor
 import co.topl.bridge.BridgeParamsDescriptor
 import co.topl.bridge.ServerConfig
 import co.topl.bridge.ToplBTCBridgeParamConfig
 import co.topl.bridge.managers.BTCWalletImpl
+import co.topl.bridge.modules.AppModule
 import co.topl.shared.utils.KeyGenerationUtils
 import org.bitcoins.keymanager.bip39.BIP39KeyManager
+import org.bitcoins.rpc.config.BitcoindAuthCredentials
 import org.http4s.ember.server.EmberServerBuilder
 import scopt.OParser
-import cats.effect.kernel.Ref
+import cats.effect.std.Queue
+import co.topl.bridge.managers.SessionEvent
 
 case class SystemGlobalState(
     currentStatus: Option[String],
@@ -19,13 +24,13 @@ case class SystemGlobalState(
     isReady: Boolean = false
 )
 
-sealed trait MintingBTCState
+sealed trait PeginSessionState
 
-case object MintingBTCState {
-  case object MintingBTCStateReady extends MintingBTCState
-  case object MintingBTCStateMinting extends MintingBTCState
-  case object MintingBTCStateWaiting extends MintingBTCState
-  case object MintingBTCStateMinted extends MintingBTCState
+case object PeginSessionState {
+  case object PeginSessionStateWaitingForBTC extends PeginSessionState
+  case class PeginSessionStateMintingTBTC(amount: Long)
+      extends PeginSessionState
+  case object PeginSessionWaitingForRedemption extends PeginSessionState
 }
 
 object Main extends IOApp with BridgeParamsDescriptor with AppModule {
@@ -36,7 +41,13 @@ object Main extends IOApp with BridgeParamsDescriptor with AppModule {
       args,
       ToplBTCBridgeParamConfig(
         toplHost = Option(System.getenv("TOPL_HOST")).getOrElse("localhost"),
-        toplWalletDb = System.getenv("TOPL_WALLET_DB")
+        toplWalletDb = System.getenv("TOPL_WALLET_DB"),
+        zmqHost = Option(System.getenv("ZMQ_HOST")).getOrElse("localhost"),
+        zmqPort =
+          Option(System.getenv("ZMQ_PORT")).map(_.toInt).getOrElse(28332),
+        btcUrl = Option(System.getenv("BTC_URL")).getOrElse("http://localhost"),
+        btcUser = Option(System.getenv("BTC_USER")).getOrElse("bitcoin"),
+        btcPassword = Option(System.getenv("BTC_PASSWORD")).getOrElse("password"),
       )
     ) match {
       case Some(config) =>
@@ -73,18 +84,55 @@ object Main extends IOApp with BridgeParamsDescriptor with AppModule {
       pegInWalletManager <- BTCWalletImpl.make[IO](pegInKm)
       walletManager <- BTCWalletImpl.make[IO](walletKm)
       logger =
-        org.typelevel.log4cats.slf4j.Slf4jLogger.getLoggerFromName[IO]("App")
+        org.typelevel.log4cats.slf4j.Slf4jLogger
+          .getLoggerFromName[IO]("btc-bridge")
       globalState <- Ref[IO].of(
         SystemGlobalState(Some("Setting up wallet..."), None)
       )
-      appAndInit <- createApp(
+      queue <- Queue.unbounded[IO, SessionEvent]
+      appAndInitAndStateMachine <- createApp(
         params,
+        "self",
+        "default",
+        queue,
         pegInWalletManager,
         walletManager,
         logger,
         globalState
       )
-      (app, init) = appAndInit
+      credentials = BitcoindAuthCredentials.PasswordBased(
+        params.btcUser,
+        params.btcPassword
+      )
+      (app, init, peginStateMachine) = appAndInitAndStateMachine
+      bitcoindInstance = BitcoinMonitor.Bitcoind.remoteConnection(
+        params.btcNetwork.btcNetwork,
+        params.btcUrl,
+        credentials
+      )
+      monitor <- BitcoinMonitor(
+        bitcoindInstance,
+        zmqHost = params.zmqHost,
+        zmqPort = params.zmqPort
+      )
+      _ <- IO.asyncForIO
+        .background(
+          fs2.Stream
+            .fromQueueUnterminated(queue)
+            .evalMap(x => peginStateMachine.innerStateConfigurer(x))
+            .compile
+            .drain
+        )
+        .allocated
+      _ <- IO.asyncForIO
+        .background(
+          monitor
+            .monitorBlocks()
+            .evalMap(x => peginStateMachine.handleNewBlock(x.block))
+            .compile
+            .drain
+        )
+        .allocated
       _ <- EmberServerBuilder
         .default[IO]
         .withIdleTimeout(ServerConfig.idleTimeOut)