+++ topl-btc-bridge/src/main/scala/co/topl/bridge/controllers/StartSessionController.scala
@@ -30,7 +30,7 @@ import java.util.UUID
 import co.topl.shared.StartPegoutSessionResponse
 import quivr.models.KeyPair
 import co.topl.shared.WalletSetupError
-import co.topl.bridge.MintingBTCState
+import co.topl.bridge.PeginSessionState
 
 object StartSessionController {
 
@@ -85,10 +85,11 @@ object StartSessionController {
         currentWalletIdx,
         mintTemplateName,
         redeemAddress,
+        address,
         scriptAsm.toHex,
         toplBridgePKey,
         sha256,
-        MintingBTCState.MintingBTCStateReady
+        PeginSessionState.PeginSessionStateWaitingForBTC
       )
     )
   }