+++ CHANGELOG.md
@@ -12,8 +12,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Minting status is now displayed in the UI.
+- Dependency on `akka-slf4j`.
+- New parameters for the launcher:  `--btc-url`, `--btc-port`, `--btc-user`, `--btc-password`, `--zmq-url`, `--zmq-port`. These allow to connect to the Bitcoin node and the ZMQ server.
+- Support for monitoring the Bitcoin network.
+
 
 ### Changed
 
 - Default port is now 4000.
 - `brambl-cli` version (for IT) is now 2.0.0-beta3.
+- Updated BramblSc to 2.0.0-beta4.
+- Integration tests to use our own Bitcoin regtest node Docker image.
+
+### Removed
+
+- `/api/confirm-deposit-btc` WS was removed.